Ai Architect/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.42.0
datasets>=2.19.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
Pillow>=10.0.0
tqdm>=4.66.0
omegaconf>=2.3.0
accelerate>=0.33.0
pyyaml>=6.0.1
wandb>=0.17.0
hydra-core>=1.3.2
